Move storage selectors into entity directory
