+++ snuba/datasets/entities/sessions.py
@@ -7,14 +7,18 @@ from snuba.clickhouse.translators.snuba.mappers import (
     ColumnToFunction,
 )
 from snuba.clickhouse.translators.snuba.mapping import TranslationMappers
+from snuba.datasets.entities.storage_selectors.selector import (
+    DefaultQueryStorageSelector,
+)
+from snuba.datasets.entities.storage_selectors.sessions import (
+    SessionsQueryStorageSelector,
+)
 from snuba.datasets.entity import Entity
 from snuba.datasets.entity_subscriptions.processors import AddColumnCondition
 from snuba.datasets.entity_subscriptions.validators import AggregationValidator
 from snuba.datasets.plans.storage_plan_builder import StorageQueryPlanBuilder
 from snuba.datasets.storage import StorageAndMappers
 from snuba.datasets.storages.factory import get_storage, get_writable_storage
-from snuba.datasets.storages.selectors.selector import DefaultQueryStorageSelector
-from snuba.datasets.storages.selectors.sessions import SessionsQueryStorageSelector
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.pipeline.simple_pipeline import SimplePipelineBuilder
 from snuba.processor import MAX_UINT32, NIL_UUID
