+++ tests/datasets/test_events.py
@@ -3,10 +3,10 @@ from snuba.clickhouse.columns import ColumnSet
 from snuba.clickhouse.translators.snuba.mapping import TranslationMappers
 from snuba.clusters.cluster import ClickhouseClientSettings
 from snuba.datasets.entities.entity_key import EntityKey
+from snuba.datasets.entities.storage_selectors.errors import ErrorsQueryStorageSelector
 from snuba.datasets.schemas.tables import TableSchema
 from snuba.datasets.storage import StorageAndMappers
 from snuba.datasets.storages.factory import get_storage, get_writable_storage
-from snuba.datasets.storages.selectors.errors import ErrorsQueryStorageSelector
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.query.data_source.simple import Entity
 from snuba.query.logical import Query
